renamed watched list test. 100% test coverage
